--- Assets/Scripts/Editor.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 65897bbab4385ab4f9e6d81d3cf15305
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
