+++ Assets/Scripts/Poached/Audio/PlaySfx.cs
@@ -1,16 +1,17 @@
-using System.Collections;
-using System.Collections.Generic;
 using UnityEngine;
 
-/// <summary>
-/// Dummy class to showcase how to use audio engine
-/// </summary>
-public class PlaySfx : MonoBehaviour
+namespace GGJ.Poached.Audio
 {
-    [SerializeField] private AudioClip clip;
-    
-    private void MyFunc()
+    /// <summary>
+    /// Dummy class to showcase how to use audio engine
+    /// </summary>
+    public class PlaySfx : MonoBehaviour
     {
-        SoundManager.Instance.PlaySound(clip);
+        [SerializeField] private AudioClip clip;
+    
+        private void MyFunc()
+        {
+            GameManager.Audio.PlaySound(clip);
+        }
     }
 }
