--- Assets/Scripts/Poached/Audio/SoundManager.cs
@@ -1,15 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using Poached.Utility;
-using UnityEngine;
-
-public class SoundManager : SingletonMonoBehaviour<SoundManager>
-{
-    [SerializeField] private AudioSource musicSource;
-    [SerializeField] private AudioSource sfxSource;
-    
-    public void PlaySound(AudioClip clip)
-    {
-        sfxSource.PlayOneShot(clip);
-    }
-}
