+++ Assets/Scripts/Poached/GameManager.cs
@@ -1,14 +1,16 @@
-using Poached.UI;
-using Poached.Utility;
+using GGJ.Poached.Audio;
+using GGJ.Poached.UI;
+using GGJ.Poached.Utility;
 using UnityEngine;
 
-namespace Poached
+namespace GGJ.Poached
 {
 [RequireComponent(typeof(PauseManager))]
+[RequireComponent(typeof(AudioManager))]
 public class GameManager : SingletonMonoBehaviour<GameManager>
 {
     // private SettingsManager _settingsManager;
-    // private AudioManager _audioManager;
+    private AudioManager _audioManager;
     // private CameraManager _cameraManager;
     private PauseManager _pauseManager;
     // private PlayerManager _playerManager;
@@ -16,7 +18,7 @@ public class GameManager : SingletonMonoBehaviour<GameManager>
     #region Public Getters
 
     // public static SettingsManager Settings => Instance._settingsManager;
-    // public static AudioManager Audio => Instance._audioManager;
+    public static AudioManager Audio => Instance._audioManager;
     // public static CameraManager Camera => Instance._cameraManager;
     public static PauseManager Pause => Instance._pauseManager;
     // public static PlayerManager Player => Instance._playerManager;
