+++ Assets/Scripts/Poached/Utility/CameraShake.cs
@@ -1,19 +1,20 @@
-﻿using System;
-using System.Collections;
+﻿using DG.Tweening;
 using UnityEngine;
-using DG.Tweening;
 
-public class CameraShake : MonoBehaviour
+namespace GGJ.Poached.Utility
 {
-    public static CameraShake Instance;
-
-    private void Awake() => Instance = this;
-
-    public static void Shake(float duration, float strength) => Instance.OnShake(duration, strength);
-
-    private void OnShake(float duration, float strength)
+    public class CameraShake : MonoBehaviour
     {
-        transform.DOShakePosition(duration, strength);
-        transform.DOShakeRotation(duration, strength);
+        public static CameraShake Instance;
+
+        private void Awake() => Instance = this;
+
+        public static void Shake(float duration, float strength) => Instance.OnShake(duration, strength);
+
+        private void OnShake(float duration, float strength)
+        {
+            transform.DOShakePosition(duration, strength);
+            transform.DOShakeRotation(duration, strength);
+        }
     }
 }
