+++ Assets/Scripts/Poached/Utility/ColorExtensions.cs
@@ -1,26 +1,28 @@
 ﻿using UnityEngine;
 
-
-public static class ColorExtensions
+namespace GGJ.Poached.Utility
 {
-    public static Color SetR(this Color color, float r)
+    public static class ColorExtensions
     {
-        return new Color(r, color.g, color.b, color.a);
-    }
+        public static Color SetR(this Color color, float r)
+        {
+            return new Color(r, color.g, color.b, color.a);
+        }
     
-    public static Color SetG(this Color color, float g)
-    {
-        return new Color(color.r, g, color.b, color.a);
-    }
+        public static Color SetG(this Color color, float g)
+        {
+            return new Color(color.r, g, color.b, color.a);
+        }
     
-    public static Color SetB(this Color color, float b)
-    {
-        return new Color(color.r, color.g, b, color.a);
-    }
+        public static Color SetB(this Color color, float b)
+        {
+            return new Color(color.r, color.g, b, color.a);
+        }
     
-    public static Color SetAlpha(this Color color, float a)
-    {
-        return new Color(color.r, color.g, color.b, a);
-    }
+        public static Color SetAlpha(this Color color, float a)
+        {
+            return new Color(color.r, color.g, color.b, a);
+        }
     
+    }
 }
