+++ Assets/Scripts/Poached/Utility/HotSwapColor.cs
@@ -1,42 +1,44 @@
-using System;
 using UnityEngine;
 
-public class HotSwapColor : MonoBehaviour
+namespace GGJ.Poached.Utility
 {
-    [SerializeField] private Color color;
-    [SerializeField] private bool hasEmission;
-    [SerializeField] private Color emissionColor;
-    [SerializeField] private float emissionIntensity;
-    [SerializeField] private MeshRenderer mr;
-    [SerializeField] private SkinnedMeshRenderer smr;
+    public class HotSwapColor : MonoBehaviour
+    {
+        [SerializeField] private Color color;
+        [SerializeField] private bool hasEmission;
+        [SerializeField] private Color emissionColor;
+        [SerializeField] private float emissionIntensity;
+        [SerializeField] private MeshRenderer mr;
+        [SerializeField] private SkinnedMeshRenderer smr;
     
-    private MaterialPropertyBlock mpb;
-    private static readonly int ShaderProp = Shader.PropertyToID("_BaseColor");
-    private static readonly int EmissionShaderProp = Shader.PropertyToID("_EmissionColor");
+        private MaterialPropertyBlock mpb;
+        private static readonly int ShaderProp = Shader.PropertyToID("_BaseColor");
+        private static readonly int EmissionShaderProp = Shader.PropertyToID("_EmissionColor");
 
-    private MaterialPropertyBlock Mpb => mpb ??= new MaterialPropertyBlock();
+        private MaterialPropertyBlock Mpb => mpb ??= new MaterialPropertyBlock();
 
-    private void OnEnable()
-    {
-        ApplyColor();
-    }
-
-    private void OnValidate()
-    {
-        ApplyColor();
-    }
-
-    private void ApplyColor()
-    {
-        Mpb.SetColor(ShaderProp, color);
-        Color finalCol = emissionColor*emissionIntensity;
-        // Remember to set the emission property to enabled in the material, and the value of emission hsv to 1
-        Mpb.SetColor(EmissionShaderProp, hasEmission ? finalCol : Color.black);
-        if(mr != null){
-            mr.SetPropertyBlock(Mpb);
+        private void OnEnable()
+        {
+            ApplyColor();
         }
-        if(smr != null){
-            smr.SetPropertyBlock(Mpb);
+
+        private void OnValidate()
+        {
+            ApplyColor();
+        }
+
+        private void ApplyColor()
+        {
+            Mpb.SetColor(ShaderProp, color);
+            Color finalCol = emissionColor*emissionIntensity;
+            // Remember to set the emission property to enabled in the material, and the value of emission hsv to 1
+            Mpb.SetColor(EmissionShaderProp, hasEmission ? finalCol : Color.black);
+            if(mr != null){
+                mr.SetPropertyBlock(Mpb);
+            }
+            if(smr != null){
+                smr.SetPropertyBlock(Mpb);
+            }
         }
     }
 }
