+++ Assets/Scripts/Poached/Utility/SingletonMonoBehaviour.cs
@@ -1,7 +1,6 @@
-﻿using System;
-using UnityEngine;
+﻿using UnityEngine;
 
-namespace Poached.Utility {
+namespace GGJ.Poached.Utility {
 public class SingletonMonoBehaviour<T> : MonoBehaviour where T : MonoBehaviour
 {
     private static T _instance;
